noJavaFx

+++ run.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
-javac -d classes */*.java
-java -cp classes Hub
+if [ -e $JAVA_HOME/jre/lib/javafx.properties ] then
+    javac -d classes */*.java
+    java -cp classes Hub
+else
+    javac -d classes */*.java
+    java -cp classes Hub noJavaFx
+fi
+
 read -p 'Clear? (y/n) ' cl
-if [ "$cl" = "y" ]
-then
+if [ "$cl" = "y" ] then
     clear
 fi
